+++ src/uint.rs
@@ -1,9 +1,11 @@
 mod public_signed;
+mod public_uint;
 mod secret_signed;
 mod secret_unsigned;
 mod traits;
 
 pub(crate) use public_signed::PublicSigned;
+pub(crate) use public_uint::PublicUint;
 pub(crate) use secret_signed::SecretSigned;
 pub(crate) use secret_unsigned::SecretUnsigned;
-pub(crate) use traits::{Exponentiable, Extendable, FromXofReader, IsInvertible, MulWide, ToMontgomery};
+pub(crate) use traits::{BoxedEncoding, Exponentiable, Extendable, FromXofReader, IsInvertible, MulWide, ToMontgomery};